+++ browser/extensions/webcompat/experiment-apis/appConstants.js
@@ -10,6 +10,9 @@ this.appConstants = class extends ExtensionAPI {
   getAPI() {
     return {
       appConstants: {
+        getAndroidPackageName: () => {
+          return Services.env.get("MOZ_ANDROID_PACKAGE_NAME");
+        },
         getEffectiveUpdateChannel: () => {
           const ver = AppConstants.MOZ_APP_VERSION_DISPLAY;
           if (ver.includes("a")) {