+++ browser/extensions/webcompat/lib/intervention_helpers.js
@@ -280,7 +280,15 @@ var InterventionHelpers = {
     },
   },
 
-  valid_platforms: ["all", "android", "desktop", "linux", "mac", "windows"],
+  valid_platforms: [
+    "all",
+    "android",
+    "desktop",
+    "fenix",
+    "linux",
+    "mac",
+    "windows",
+  ],
   valid_channels: ["beta", "esr", "nightly", "stable"],
 
   shouldSkip(intervention, firefoxVersion, firefoxChannel) {
@@ -329,6 +337,12 @@ var InterventionHelpers = {
         platformInfo.os,
         platformInfo.os == "android" ? "android" : "desktop",
       ];
+      if (platformInfo.os == "android") {
+        const packageName = await browser.appConstants.getAndroidPackageName();
+        if (packageName.includes("fenix") || packageName.includes("firefox")) {
+          InterventionHelpers._platformMatches.push("fenix");
+        }
+      }
     }
     return InterventionHelpers._platformMatches;
   },