+++ dom/base/ChromeUtils.h
@@ -329,6 +329,8 @@ class ChromeUtils {
   static already_AddRefed<Promise> GetGMPContentDecryptionModuleInformation(
       GlobalObject& aGlobal, ErrorResult& aRv);
 
+  static void AndroidMoveTaskToBack(GlobalObject& aGlobal);
+
  private:
   // Number of DevTools session debugging the current process
   static std::atomic<uint32_t> sDevToolsOpenedCount;